fixed compiling error
there was an error when trying to import. required export and i fixed that. then modules needed to be created by adding an export to one of the files. re named one of the files as well since it was redundant. pushing production build after this hopefully to have a static demo to display functionality.